requirements.txt
# Python=3.10.4
Django>=4.2.1,<4.3
djangorestframework>=3.14.0,<3.15
psycopg>=3.1.9,<3.2
drf-spectacular>=0.26.2,<0.27
drf-spectacular-sidecar>=2023.5.1,<2024.6
djangorestframework-simplejwt>=5.2.2,<5.3
flake8>=6.0.0,<6.1
django-cors-headers>=4.0.0,<4.1
black>=23.3.0,<23.4 # black is a code formatter
Pillow>=9.5.0,<9.6 # for image processing

